# packages/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Montessori ERP & LMS API
  description: |
    Multi-tenant, role-based ERP & Learning Management System for Montessori schools.
    
    ## Authentication
    All routes (except `/api/v1/auth/*`) require a JWT Bearer token in the `Authorization` header.
    
    ## Roles
    - **admin**: Full access to all modules
    - **teacher**: Students, attendance, observations, curriculum, messages
    - **parent**: Read-only access to own children, fees, messages
    - **student**: Read-only curriculum and announcements
  version: 1.0.0
  contact:
    name: Montessori ERP Support
servers:
  - url: http://localhost:4001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/students:
    get:
      tags: [Students]
      summary: List students
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: classId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated students list
    post:
      tags: [Students]
      summary: Create a student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, dob]
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                dob:
                  type: string
                  format: date
                gender:
                  type: string
                classId:
                  type: string
      responses:
        '201':
          description: Student created
  /api/v1/students/{id}:
    get:
      tags: [Students]
      summary: Get student by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student details
        '404':
          description: Student not found
    patch:
      tags: [Students]
      summary: Update student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                classId:
                  type: string
      responses:
        '200':
          description: Student updated
    delete:
      tags: [Students]
      summary: Delete student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Student deleted

  # ==================== CLASSES ====================
...
  /api/v1/observations:
    get:
      tags: [Observations]
      summary: List observations
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Observations list
    post:
      tags: [Observations]
      summary: Create observation (teacher only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, note, masteryLevel]
              properties:
                studentId:
                  type: string
                curriculumItemId:
                  type: string
                note:
                  type: string
                masteryLevel:
                  type: string
                  enum: [INTRODUCED, PRACTICING, MASTERED]
      responses:
        '201':
          description: Observation created
  /api/v1/observations/{id}:
    patch:
      tags: [Observations]
      summary: Update observation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Observation updated
    delete:
      tags: [Observations]
      summary: Delete observation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Observation deleted
  /api/v1/observations/student/{studentId}/progress:
    get:
      tags: [Observations]
      summary: Get student progress from observations
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student progress data

  # ==================== CURRICULUM ====================
...
  /api/v1/curriculum/progress/student/{studentId}:
    get:
      tags: [Curriculum]
      summary: Get student's curriculum progress
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student progress

  # ==================== FINANCE ====================
...
  /api/v1/finance/invoices/student/{studentId}:
    get:
      tags: [Finance]
      summary: Get invoices for a student
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student invoices
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List materials
      security:
        - bearerAuth: []
      parameters:
        - name: lowStock
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        '200':
          description: Materials list
    post:
      tags: [Inventory]
      summary: Add material
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                quantity:
                  type: integer
                location:
                  type: string
                lowStockThreshold:
                  type: integer
      responses:
        '201':
          description: Material created
  /api/v1/inventory/{id}:
    get:
      tags: [Inventory]
      summary: Get material
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Material details
    patch:
      tags: [Inventory]
      summary: Update material
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Material updated
    delete:
      tags: [Inventory]
      summary: Delete material
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Material deleted
  /api/v1/inventory/{id}/stock:
    patch:
      tags: [Inventory]
      summary: Adjust stock level
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adjustment]
              properties:
                adjustment:
                  type: integer
                  description: Positive to add, negative to remove
      responses:
        '200':
          description: Stock adjusted
  /api/v1/inventory/checkout:
    post:
      tags: [Inventory]
      summary: Checkout material to classroom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [materialId]
              properties:
                materialId:
                  type: string
                classId:
                  type: string
                quantity:
                  type: integer
                  default: 1
      responses:
        '201':
          description: Checkout recorded
  /api/v1/inventory/checkout/{id}/return:
    post:
      tags: [Inventory]
      summary: Return checked-out material
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Material returned

  # ==================== COMMUNICATION ====================
...
  /api/v1/gamification/points/student/{studentId}:
    get:
      tags: [Gamification]
      summary: Get student points
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student points
...
  /api/v1/ai/insights/student/{studentId}:
    get:
      tags: [AI]
      summary: Get AI insights for a student (teacher/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: AI-generated student insights
        '403':
          description: Insufficient permissions
...